backend/requirements.txt
# FastAPI Framework - Modern web framework for building APIs with Python
fastapi==0.104.1

# Uvicorn - ASGI server for running FastAPI applications
uvicorn==0.24.0

# SQLAlchemy - SQL toolkit and Object-Relational Mapping (ORM) for database operations
sqlalchemy==2.0.23

# Psycopg2 - PostgreSQL database adapter for Python
psycopg2-binary==2.9.9

# Python-dotenv - Load environment variables from .env files
python-dotenv==1.0.0

# Pydantic - Data validation and settings management using Python type annotations
pydantic==2.5.0
pydantic-settings==2.1.0

# PyJWT - JWT token encoding and decoding for authentication
pyjwt==2.8.1

# Passlib - Password hashing library for secure credential storage
passlib==1.7.4

# Python-bcrypt - Cryptographic library for password hashing
bcrypt==4.1.1

# Requests - HTTP library for making API calls
requests==2.31.0

# CORS - Cross-Origin Resource Sharing support for FastAPI
python-multipart==0.0.6

# OpenAI - For AI/Whisper integration (optional)
openai==1.3.5

# Pillow - Image processing library
Pillow==10.1.0

# Alembic - Database migration tool for SQLAlchemy
alembic==1.13.0

# Python-jose - JWT token handling
python-jose==3.3.0

# Email validation
email-validator==2.1.0
